--- components/date-picker/demo/bordered.md
@@ -1,7 +1,0 @@
-## zh-CN
-
-无边框样式。
-
-## en-US
-
-Bordered-less style component.
