+++ components/date-picker/demo/variant.md
@@ -1,0 +1,7 @@
+## zh-CN
+
+可选 `outlined` `filled` `borderless` 三种形态。
+
+## en-US
+
+There are `outlined` `fille` and `borderless`, totally three variants to choose from.
