+++ components/date-picker/demo/variant.tsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import { DatePicker, Flex } from 'antd';
+
+const { RangePicker } = DatePicker;
+
+const App: React.FC = () => (
+  <Flex vertical gap={12}>
+    <Flex gap={8}>
+      <DatePicker placeholder="Outlined" />
+      <RangePicker placeholder={['Outlined', '']} />
+    </Flex>
+    <Flex gap={8}>
+      <DatePicker placeholder="Filled" variant="filled" />
+      <RangePicker placeholder={['Filled', '']} variant="filled" />
+    </Flex>
+    <Flex gap={8}>
+      <DatePicker placeholder="Borderless" variant="borderless" />
+      <RangePicker placeholder={['Borderless', '']} variant="borderless" />
+    </Flex>
+  </Flex>
+);
+
+export default App;
