+++ components/date-picker/generatePicker/generateRangePicker.tsx
@@ -29,6 +29,8 @@ import Components from './Components';
 import type { CommonPickerMethods, PickerComponentClass } from './interface';
 import { useZIndex } from '../../_util/hooks/useZIndex';
 import useCSSVarCls from '../../config-provider/hooks/useCSSVarCls';
+import useVariant from '../../_util/hooks/useVariants';
+import { InputVariants } from '../../input/Input';
 
 export default function generateRangePicker<DateType>(generateConfig: GenerateConfig<DateType>) {
   type InternalRangePickerProps = RangePickerProps<DateType> & {};
@@ -62,6 +64,7 @@ export default function generateRangePicker<DateType>(generateConfig: GenerateCo
       clearIcon,
       allowClear,
       rootClassName,
+      variant: customVariant,
       ...restProps
     } = props;
 
@@ -71,6 +74,8 @@ export default function generateRangePicker<DateType>(generateConfig: GenerateCo
     const { compactSize, compactItemClassnames } = useCompactItemContext(prefixCls, direction);
     const { format, showTime, picker } = props as any;
     const rootPrefixCls = getPrefixCls();
+
+    const [variant, enableVariantCls] = useVariant(customVariant, bordered, InputVariants);
 
     const rootCls = useCSSVarCls(prefixCls);
     const [wrapCSSVar, hashId, cssVarCls] = useStyle(prefixCls, rootCls);
@@ -85,6 +90,8 @@ export default function generateRangePicker<DateType>(generateConfig: GenerateCo
       const warning = devUseWarning('DatePicker.RangePicker');
 
       warning.deprecated(!dropdownClassName, 'dropdownClassName', 'popupClassName');
+
+      warning.deprecated(!('bordered' in props), 'bordered', 'variant');
     }
 
     // ===================== Size =====================
@@ -139,7 +146,7 @@ export default function generateRangePicker<DateType>(generateConfig: GenerateCo
         className={classNames(
           {
             [`${prefixCls}-${mergedSize}`]: mergedSize,
-            [`${prefixCls}-borderless`]: !bordered,
+            [`${prefixCls}-${variant}`]: enableVariantCls,
           },
           getStatusClassNames(prefixCls, getMergedStatus(contextStatus, customStatus), hasFeedback),
           hashId,
