+++ components/date-picker/style/variants.ts
@@ -1,0 +1,13 @@
+import type { PickerToken } from './token';
+import type { CSSObject } from '@ant-design/cssinjs';
+import { genBorderlessStyle, genFilledStyle, genOutlinedStyle } from '../../input/style/variants';
+
+const genVariantsStyle = (token: PickerToken): CSSObject => ({
+  [token.componentCls]: {
+    ...genOutlinedStyle(token),
+    ...genFilledStyle(token),
+    ...genBorderlessStyle(token),
+  },
+});
+
+export default genVariantsStyle;
